final commit i promise

+++ routes/shoppingCart.js
@@ -19,7 +19,6 @@ router.get("/", async (req, res) => {
 
     const cart = await Shopping_cart.findAll({ where: { buyer_id: id } });
 
-    console.log(cart.length);
     if (cart.length == 0)
       return res.status(404).send("Error : User's cart is empty");
 
@@ -45,7 +44,7 @@ router.post("/", async (req, res) => {
 
   const buyer = await User.findOne({ where: { user_id: buyer_id } });
 
-  if (!buyer) return res.send("Error : Buyer not found ").status(404);
+  if (!buyer) return res.status(404).send("Error : Buyer not found ");
 
   for (let productToAdd of products) {
     var product = await Product.findOne({
@@ -72,7 +71,6 @@ router.post("/", async (req, res) => {
         .send(
           `Error : Seller for product with id=${productToAdd.product_id} not found`
         );
-    //------------------------------
 
     const stock = await Stock.findOne({
       where: {
@@ -84,14 +82,7 @@ router.post("/", async (req, res) => {
       return res.status(400).send(e.message);
     });
     if (!stock) return res.send.status(404).send("Error : Stock not found ");
-    //------------------------------
 
-    /*     stock.quantity = stock.quantity - productToAdd.quantity;
-    stock.save();
-    product.stock = product.stock - productToAdd.quantity;
-    product.save();
- */
-    //------------------------------
     if (stock.quantity < productToAdd.quantity)
       return res
         .status(400)
@@ -114,15 +105,6 @@ router.post("/", async (req, res) => {
       overwrite.quantity = productToAdd.quantity;
       overwrite.product = product;
       overwrite.save();
-
-      //------------------------------
-
-      /*       stock.quantity = stock.quantity + productToAdd.quantity;
-      stock.save();
-      product.stock = product.stock + productToAdd.quantity;
-      product.save(); */
-
-      //------------------------------
 
       result.push(overwrite);
     } else {
@@ -162,16 +144,10 @@ router.post("/", async (req, res) => {
 //--------------------------------------------------------------------------------------------
 
 router.delete("/", async (req, res) => {
-  const { buyer_id, seller_id, product_id } = req.body;
+  const { buyer_id, seller_id, target } = req.body;
 
-  if (!buyer_id || !seller_id || !product_id)
+  if (!buyer_id || !seller_id || !target)
     return res.status(400).send("Error : Missing data in request");
-  const product = await Product.findOne({
-    where: {
-      product_id: product_id,
-    },
-  });
-  if (!product) return res.status(404).send("Error : Product not found");
   const buyer = await User.findOne({
     where: {
       user_id: buyer_id,
@@ -185,33 +161,50 @@ router.delete("/", async (req, res) => {
   });
   if (!seller) return res.status(404).send("Error : Seller not found");
 
-  const cart = await Shopping_cart.findOne({
-    where: {
-      buyer_id: buyer_id,
-      seller_id: seller_id,
-      product_id: product_id,
-    },
-  });
-
-  if (!cart) return res.status(404).send("Error : Item not found on cart");
-
-  Shopping_cart.destroy({
-    where: {
-      buyer_id: buyer_id,
-      seller_id: seller_id,
-      product_id: product_id,
-    },
-  })
-    .then((rowsDeleted) => {
-      if (rowsDeleted === 1) {
-        return res.status(200).send("Deleted succesfully");
-      }
-      return res.status(400).send("Delete unsuccesfull");
-    })
-    .catch((e) => {
-      console.log(e);
-      return res.status(400).send(e.message);
+  if (target == "ALL") {
+    Shopping_cart.destroy({
+      where: {
+        user_id: buyer_id,
+      },
+    }).then(() => {
+      return res.status(200).send("User cart deleted");
     });
+  } else {
+    const product = await Product.findOne({
+      where: {
+        product_id: target,
+      },
+    });
+    if (!product) return res.status(404).send("Error : Product not found");
+
+    const cart = await Shopping_cart.findOne({
+      where: {
+        buyer_id: buyer_id,
+        seller_id: seller_id,
+        product_id: target,
+      },
+    });
+
+    if (!cart) return res.status(404).send("Error : Item not found on cart");
+
+    Shopping_cart.destroy({
+      where: {
+        buyer_id: buyer_id,
+        seller_id: seller_id,
+        product_id: target,
+      },
+    })
+      .then((rowsDeleted) => {
+        if (rowsDeleted === 1) {
+          return res.status(200).send("Deleted succesfully");
+        }
+        return res.status(400).send("Delete unsuccesfull");
+      })
+      .catch((e) => {
+        console.log(e);
+        return res.status(400).send(e.message);
+      });
+  }
 });
 
 //--------------------------------------------------------------------------------------------
